refactor: extract super ops support actions
